feat: change some Controller and Commands

+++ database/migrations/2025_03_23_180943_create_fuel_properties_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('fuel_type');
             $table->string('unit');
-            $table->float('conversion_factor');
+            $table->decimal('conversion_factor', 15, 10);
             $table->float('co2_factor');
             $table->float('ch4_factor');
             $table->float('n2o_factor');
